requirements.txt
flask==3.0.3
flask-restful==0.3.10
google-cloud-bigquery==3.25.0
pytest==8.1.1
pytest-mock==3.14.0
pytest-flask==1.3.0
requests-mock==1.12.1
